pytest: Migrate `env` to pyln-testing
